Remove date extension and move code into blog.js
